examples_e2e/fianancial_transcripts/requirements.txt
azure-ai-formrecognizer
psycopg2
python-dotenv
openai
pandas
numpy
pgvector
langchain
docx2txt
docx2pdf
tiktoken
scikit-learn
mlflow
azure-storage-blob
